% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_ROILabelVars.R
\name{Add_ROILabelVars}
\alias{Add_ROILabelVars}
\title{Add Variables to SizeExractR Database, based on ROI Label Translator Matrix}
\usage{
Add_ROILabelVars(data, label.translator)
}
\arguments{
\item{data}{An object of class dataframe as output directly from SizeExtractR::Build_Uncalibrated_Dataset()}

\item{label.translator}{An object of class data.frame as output directly from SizeExtractR::CheckSet_ROILabelVars. See documentation for that function for more information.}
}
\value{
Adds to the SizeExtractR database (and returns this dataframe object) by deconstructing the ROI.Label variable and forming it into new columns showing TRUE or FALSE (i.e. yes or no) for each of the different labelling variables.

Note that the labelling variables are for ROIs of interest, not the calibration ROIs (ROI.Code = "Cali_Pts" or "M" denoting calibration points and measurement length respectively).
}
\description{
Add Variables to SizeExractR Database, based on ROI Label Translator Matrix
}
\examples{
Database.ROILab = Add_ROILabelVars(Database, Label.Translator)

}
