% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibrate_Database.R
\name{Calibrate_Database}
\alias{Calibrate_Database}
\title{Calibrate the SizeExtractR database, based on known calibration length (cm)}
\usage{
Calibrate_Database(datalab, known.length)
}
\arguments{
\item{known.length}{a numerical value of the length of calibration lengths from image analyses (in centimeters).}

\item{data}{An object of class dataframe as output directly from SizeExtractR::Add_ROILabelVars()}
}
\value{
For each image independently, this function returns a calibrated  SizeExtractR database (dataframe object).

It does these calibrations on a per image basis. The mean number of pixels is calculated among all the calibration lengths in that image (i.e., ROIs with ROI.Code of "M"). This is then compared to the 'known.length' parameter, to compute all aspects of size of each ROI.

Length, Position, Area, and Volume are all given in square centimeters.
}
\description{
Calibrate the SizeExtractR database, based on known calibration length (cm)
}
\examples{
Database.cal = Calibrate_Database(Database.ROILab, known.length = 10)

}
