% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_ROI_Types.R
\name{Check_ROI_Types}
\alias{Check_ROI_Types}
\title{Check for mistakes in ROI Types from the SizeExctractR ImageJ tools and protocol}
\usage{
Check_ROI_Types(data)
}
\arguments{
\item{data}{An object of class dataframe as output directly from SizeExtractR::Build_Uncalibrated_Dataset()}
}
\value{
Runs through interactive checks for the user to determine if the ROI Types (i.e., main type of each object i.e. Calibration point, calibration length or ROI of interest) used in their image analysis are correct, and help to identify any images that had mistakes in the labelling process. If mistakes are found support information is given to the user.

Also returns a character string reading "success", which is not used in further SizeExtractR functions.
}
\description{
Check for mistakes in ROI Types from the SizeExctractR ImageJ tools and protocol
}
\examples{
# load in the output of Build_Uncalibrated_Dataset
data(Database)

# Run the function
ROI.Types = Check_ROI_Types(Database)
if(length(ROI.Types) > 0){print(ROI.Types)}
}
