% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckSet_ROILabelVars.R
\name{CheckSet_ROILabelVars}
\alias{CheckSet_ROILabelVars}
\title{Check for mistakes in ROI Labels from the SizeExctractR ImageJ tools and protocol}
\usage{
CheckSet_ROILabelVars(data, path)
}
\arguments{
\item{data}{An object of class dataframe as output directly from SizeExtractR::Build_Uncalibrated_Dataset()}

\item{path}{A Directory path within which holds the text files outputted from SizeExtractR imageJ tools and protocol. The directory path should be given in double quotes in the same format as that returned by the function getwd(), and should be the same as that used in SizeExtractR::Build_Uncalibrated_Dataset()}
}
\value{
Runs through interactive checks for the user to determine if the ROI Labels (i.e., additional notes about each ROI of interest) used in their image analysis are correct, and help to identify any images that had mistakes in the labelling process. If mistakes are found support information is given to the user.

Prompts the user to set up a Label csv file with the required information needed to create the label translator matix

Once the csv file has been created and edited propoerly, this function returns a Label Translator Matrix. This can be used to check that the Label codes will translent to the correct Variable names.
}
\description{
Check for mistakes in ROI Labels from the SizeExctractR ImageJ tools and protocol
}
\examples{
# load in the output of Build_Uncalibrated_Dataset
mypath = paste0(.libPaths()[1],"/SizeExtractR/TextFiles")
data(Database)

# Run the function
Label.Translator = CheckSet_ROILabelVars(Database, mypath)

if(length(Label.Translator) > 0){
  print(Label.Translator, row.names = F)
}

}
