% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.cal.R
\docType{data}
\name{Database.cal}
\alias{Database.cal}
\title{Calibrated Dataframe}
\format{
An object of class dataframe
}
\usage{
data(Database.cal)
}
\description{
A dataframe of the example data provided with the package. Shown once Area Metrics are calibrated using SizeExtractR::Calibrate_Database().
}
\examples{
head(Database.cal)
}
\keyword{datasets}
