% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.groups.R
\docType{data}
\name{Database.groups}
\alias{Database.groups}
\title{Calibrated Dataframe with Three Grouping Variables}
\format{
An object of class dataframe
}
\usage{
data(Database.groups)
}
\description{
A dataframe of an extended dataset as that provided with the package for a subtropical coral Pocillopora aliciae, with three grouping variables: Year, Site, and Transect. Data from Lachs et al. (2021) Coral Reefs.
}
\examples{
head(Database.groups)
}
\keyword{datasets}
