% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\docType{data}
\name{Database}
\alias{Database}
\title{Dataframe after initial build}
\format{
An object of class dataframe
}
\usage{
data(Database)
}
\description{
A dataframe of the example data provided with the package. Shown once initally built using SizeExtractR::Build_Uncalibrated_Dataset().
}
\examples{
head(Database)
}
\keyword{datasets}
