% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.ROILab.R
\docType{data}
\name{Database.ROILab}
\alias{Database.ROILab}
\title{Dataframe with labels converted to variables}
\format{
An object of class dataframe
}
\usage{
data(Database.ROILab)
}
\description{
A dataframe of the example data provided with the package. SHown once ROI labels are converted to Variables using SizeExtractR::Add_ROILabelVars().
}
\examples{
head(Database.ROILab)
}
\keyword{datasets}
