% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.ROILabelCode.R
\docType{data}
\name{Database.ROILabelCode}
\alias{Database.ROILabelCode}
\title{Dataframe with ROI label codes converted to variables}
\format{
An object of class dataframe
}
\usage{
data(Database.ROILabelCode)
}
\description{
A dataframe of the example data provided with the package. SHown once ROI label Codes are converted to Variables using SizeExtractR::Add_ROILabelCodeVars().
}
\examples{
head(Database.ROILabelCode)
}
\keyword{datasets}
