% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Full_SizeExtractR_Workflow.R
\name{Full_SizeExtractR_Workflow}
\alias{Full_SizeExtractR_Workflow}
\title{Run the full SizeExtractR workflow - a series of nested functions to extract all sizes.}
\usage{
Full_SizeExtractR_Workflow(
  path,
  known.calibration.length,
  include.calibrations
)
}
\arguments{
\item{path}{A Directory path within which holds the text files outputted from SizeExtractR imageJ tools and protocol. The directory path should be given in double quotes in the same format as that returned by the function getwd()}

\item{known.calibration.length}{a numerical value of the length of calibration lengths from image analyses (in centimeters).}

\item{include.calibrations}{either TRUE of FALSE, determines whether or not to include the calibration-length ROIs alongside the ROIs of interest.}
}
\value{
Returns a calibrated dataset of size metrics, that includes user defined variables based on 1) the directory structure, and 2) the ROI labels used in image analyses.

All relevant interactive checks are also undertaken to determine whether the exported database is correct or not.
}
\description{
Run the full SizeExtractR workflow - a series of nested functions to extract all sizes.
}
\examples{
# Set directory within which all text files are located
mypath = paste0(path.package("SizeExtractR"), "/inst/TextFiles")

# Run the function

Database.cal.SizeOnly = Full_SizeExtractR_Workflow(mypath, 1, FALSE)

# Show dataset
if(length(Database.cal.SizeOnly) > 0){
  head(Database.cal.SizeOnly)
  colnames(Database.cal.SizeOnly)
  hist(Database.cal.SizeOnly$Area, xlab = bquote(Egg~Area~(cm^2)), main = "Size Only Dataset")
}

}
