% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Label.Translator.R
\docType{data}
\name{Label.Translator}
\alias{Label.Translator}
\title{Label Translator Matrix}
\format{
An object of class dataframe
}
\usage{
data(Label.Translator)
}
\description{
A dataframe showing the translation from the coded labels attached to Regions of Interest (ROI) in imageJ, to dataframe varibales (TRUE/FALSE).
}
\examples{
head(Label.Translator)
}
\keyword{datasets}
