% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SizeOnly_Database.R
\name{SizeOnly_Database}
\alias{SizeOnly_Database}
\title{Reduce the calibrated SzeExtractR database by removing all calibration ROIs}
\usage{
SizeOnly_Database(datacal)
}
\arguments{
\item{datacal}{An object of class dataframe as output directly from SizeExtractR::Calibrate_Database()}
}
\value{
Returns a calibrated SizeExtractR database (i.e., dataframe object) without points and lengths used in calibrations.
}
\description{
Reduce the calibrated SzeExtractR database by removing all calibration ROIs
}
\examples{
Database.SizeOnly = SizeOnly_Database(Database.cal)

}
