% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varnames.R
\docType{data}
\name{varnames}
\alias{varnames}
\title{Variables Names from Directory Structure}
\format{
An object of class vector
}
\usage{
data(varnames)
}
\description{
A vector of Variable Names that will be used later in database formation within SizeExtractR.
}
\examples{
head(varnames)
}
\keyword{datasets}
