% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckSet_DirecVars.R
\name{CheckSet_DirecVars}
\alias{CheckSet_DirecVars}
\title{Check Directory Structure, and Set Directory-derived Variable Names}
\usage{
CheckSet_DirecVars(path)
}
\arguments{
\item{path}{A Directory path within which holds the text files outputted from SizeExtractR imageJ tools and protocol. The directory path should be given in double quotes in the same format as that returned by the function getwd()}
}
\value{
Returns a vector of Variable Names that will be used later in database formation within SizeExtractR.

It also provides a user interface to check the Directory structure is suitable for other SizeExtractR functions. If the Directory structure is incorrect this function prompts the user about how to fix the directory structure.
}
\description{
Check Directory Structure, and Set Directory-derived Variable Names
}
\examples{
# Set the path within which all text files (from image analysis) are found
mypath = paste0(.libPaths()[1],"/SizeExtractR/TextFiles")

# Run the function
varnames = CheckSet_DirecVars(mypath)
if(length(varnames) > 0){print(varnames)}

}
