% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SizeOnly_Database.R
\name{SizeOnly_Database}
\alias{SizeOnly_Database}
\title{Reduce the calibrated SzeExtractR database by removing all calibration ROIs}
\usage{
SizeOnly_Database(datacal)
}
\arguments{
\item{datacal}{An object of class dataframe as output directly from SizeExtractR::Calibrate_Database()}
}
\value{
Returns a calibrated SizeExtractR database (i.e., dataframe object) without points and lengths used in calibrations. THis is helpful when it comes to indexing the dataset for plotting using ggplot or other packages.
}
\description{
Reduce the calibrated SzeExtractR database by removing all calibration ROIs
}
\examples{
# load in the output of Add_ROILabelVars
data(Database.cal)

# Run the function
Database.SizeOnly = SizeOnly_Database(Database.cal)

# Histogram of Egg Sizes - Note the indexing is not needed for the latter dataset
par(mfrow = c(1,2))

# Full Dataset
ind=which(Database.cal$ROI.Type == "e")
hist(Database.cal$Area[ind], xlab = bquote(Egg~Area~(cm^2)), main = "Full Dataset")

# SizeOnly Dataset
hist(Database.SizeOnly$Area, xlab = bquote(Egg~Area~(cm^2)), main = "Size Only Dataset")

}
